pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kanban-prompt-companion"
version = "0.1.0"
description = "Local-first voice-note to Kanban prompt package companion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "pydantic>=2.8.0",
  "python-frontmatter>=1.1.0",
  "pytest>=8.0.0",
  "watchdog>=4.0.1",
  "uvicorn[standard]>=0.30.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
